Move shared types into a targetSupply folder
