+++ src/schema/v2/types/targetSupply/recentlySoldArtworksConnection.ts
